requirements.txt
matplotlib
numpy
Pillow
pytorch-lightning
pycocotools
timm
torch
torchvision
transformers
wandbclear